+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Document how to work with fork process web server models(Gunicorn, uWSGI etc...)
   ([#1609](https://github.com/open-telemetry/opentelemetry-python/pull/1609))
+### Changed
+- Rename `IdsGenerator` to `IdGenerator`
+  ([#1651])(https://github.com/open-telemetry/opentelemetry-python/pull/1651)
 
 ## [0.18b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v0.18b0) - 2021-02-16